requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
slowapi==0.1.9
structlog==25.1.0
pydantic-settings==2.9.1
cachetools==5.5.2
